Refactor visual, nuevas vistas y controlador
